pyproject.toml
[project]
name = "self-forcing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "accelerate==1.9.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "antlr4-python3-runtime==4.9.3",
    "anyio==4.9.0",
    "appdirs==1.4.4",
    "argon2-cffi==25.1.0",
    "argon2-cffi-bindings==25.1.0",
    "arrow==1.3.0",
    "asttokens==3.0.0",
    "async-lru==2.0.5",
    "async-timeout==5.0.1",
    "attrs==25.3.0",
    "autobahn==24.4.2",
    "automat==25.4.16",
    "av==13.1.0",
    "babel==2.17.0",
    "beautifulsoup4==4.13.4",
    "bidict==0.23.1",
    "bleach==6.2.0",
    "blinker==1.9.0",
    "certifi==2025.7.14",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "clip",
    "coloredlogs==15.0.1",
    "comm==0.2.3",
    "constantly==23.10.4",
    "contourpy==1.3.2",
    "cryptography==45.0.5",
    "cycler==0.12.1",
    "dashscope==1.24.0",
    "debugpy==1.8.16",
    "decorator==5.2.1",
    "defusedxml==0.7.1",
    "diffusers==0.31.0",
    "distro==1.9.0",
    "dominate==2.9.1",
    "easydict==1.13",
    "einops==0.8.1",
    "exceptiongroup==1.3.0",
    "executing==2.2.0",
    "fal>=1.38.0",
    "fastapi==0.116.1",
    "fastjsonschema==2.21.1",
    "filelock==3.18.0",
    "flask==3.1.1",
    "flask-socketio==5.5.1",
    "flatbuffers==25.2.10",
    "fonttools==4.59.0",
    "fqdn==1.5.1",
    "frozenlist==1.7.0",
    "fsspec==2025.7.0",
    "ftfy==6.3.1",
    "gitdb==4.0.12",
    "gitpython==3.1.45",
    "h11==0.16.0",
    "hf-xet==1.1.5",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "huggingface-hub==0.34.3",
    "humanfriendly==10.0",
    "humanize==4.12.3",
    "hyperlink==21.0.0",
    "idna==3.10",
    "imageio==2.37.0",
    "imageio-ffmpeg==0.6.0",
    "importlib-metadata==8.7.0",
    "incremental==24.7.2",
    "inquirerpy==0.3.4",
    "ipykernel==6.30.1",
    "ipython==8.37.0",
    "ipywidgets==8.1.7",
    "isoduration==20.11.0",
    "iterable-io==1.0.0",
    "itsdangerous==2.2.0",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "json5==0.12.1",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.0",
    "jsonschema-specifications==2025.4.1",
    "jupyter==1.1.1",
    "jupyter-client==8.6.3",
    "jupyter-console==6.6.3",
    "jupyter-core==5.8.1",
    "jupyter-events==0.12.0",
    "jupyter-lsp==2.2.6",
    "jupyter-server==2.16.0",
    "jupyter-server-terminals==0.5.3",
    "jupyterlab==4.4.5",
    "jupyterlab-pygments==0.3.0",
    "jupyterlab-server==2.27.3",
    "jupyterlab-widgets==3.0.15",
    "kiwisolver==1.4.8",
    "lark==1.2.2",
    "lazy-loader==0.4",
    "lmdb==1.7.3",
    "magic-wormhole==0.20.0",
    "mako==1.3.10",
    "markupsafe==3.0.2",
    "matplotlib==3.10.3",
    "matplotlib-inline==0.1.7",
    "mistune==3.1.3",
    "ml-dtypes==0.5.3",
    "mpmath==1.3.0",
    "msgpack==1.1.1",
    "multidict==6.6.3",
    "nbclient==0.10.2",
    "nbconvert==7.16.6",
    "nbformat==5.10.4",
    "nest-asyncio==1.6.0",
    "networkx==3.4.2",
    "ninja==1.13.0",
    "notebook==7.4.5",
    "notebook-shim==0.2.4",
    "numpy==2.1.2",
    "omegaconf==2.3.0",
    "open-clip-torch==3.0.0",
    "openai==1.100.1",
    "opencv-python==4.11.0.86",
    "overrides==7.7.0",
    "packaging==25.0",
    "pandocfilters==1.5.1",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "pfzy==0.3.4",
    "pillow==10.4.0",
    "pip==25.2",
    "platformdirs==4.3.8",
    "prometheus-client==0.22.1",
    "prompt-toolkit==3.0.51",
    "propcache==0.3.2",
    "protobuf==6.31.1",
    "psutil==7.0.0",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycocotools==2.0.10",
    "pycparser==2.22",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pygments==2.19.2",
    "pynacl==1.5.0",
    "pyopenssl==25.1.0",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-engineio==4.12.2",
    "python-json-logger==3.3.0",
    "python-socketio==5.13.0",
    "pytools==2025.2.2",
    "pyyaml==6.0.2",
    "pyzmq==27.0.1",
    "qrcode==8.2",
    "redis[hiredis]>=6.4.0",
    "referencing==0.36.2",
    "regex==2025.7.34",
    "requests==2.32.4",
    "rfc3339-validator==0.1.4",
    "rfc3986-validator==0.1.1",
    "rfc3987-syntax==1.1.0",
    "rpds-py==0.27.0",
    "safetensors==0.5.3",
    "scikit-image==0.25.2",
    "scipy==1.15.3",
    "send2trash==1.8.3",
    "sentencepiece==0.2.0",
    "sentry-sdk==2.34.1",
    "service-identity==24.2.0",
    "setuptools==80.9.0",
    "simple-websocket==1.1.0",
    "siphash24==1.7",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "spake2==0.9",
    "stack-data==0.6.3",
    "starlette==0.47.2",
    "sympy==1.14.0",
    "tensorboard>=2.20.0",
    "terminado==0.18.1",
    "tifffile==2025.5.10",
    "timm==1.0.19",
    "tinycss2==1.4.0",
    "tokenizers==0.21.4",
    "tomli==2.2.1",
    "torch==2.8.0",
    "torch-tb-profiler>=0.4.3",
    "torchao==0.12.0",
    "torchaudio==2.8.0",
    "torchvision==0.23.0",
    "tornado==6.5.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "transformers==4.54.1",
    "triton==3.4.0",
    "twisted==25.5.0",
    "txaio==25.6.1",
    "txtorcon==24.8.0",
    "types-python-dateutil==2.9.0.20250809",
    "typing-extensions==4.14.1",
    "uri-template==1.3.0",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "wandb==0.21.0",
    "wcwidth==0.2.13",
    "webcolors==24.11.1",
    "webencodings==0.5.1",
    "websocket-client==1.8.0",
    "werkzeug==3.1.3",
    "widgetsnbextension==4.0.14",
    "wsproto==1.2.0",
    "yarl==1.20.1",
    "zipp==3.23.0",
    "zipstream-ng==1.8.0",
    "zope-interface==7.2",
]

[tool.uv.extra-build-dependencies]
nvidia-pyindex = ["pip"]

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git", rev = "dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1" }

[dependency-groups]
dev = [
    "line-profiler>=5.0.0",
    "tensorboard>=2.20.0",
    "torch-tb-profiler>=0.4.3",
]
